+++ apps/hubber/src/main.ts
@@ -15,7 +15,7 @@ logger.info(`Klave Hubber API v${process.env.GIT_REPO_VERSION}`);
 logger.info(`Branch ${process.env.GIT_REPO_BRANCH} - ${process.env.GIT_REPO_COMMIT?.substring(0, 8)}${process.env.GIT_REPO_DIRTY ? '*' : ''}`);
 logger.info(`Node ${process.version} - ${process.cwd()}`);
 
-dbOps.initialize()
+const serverHandle = dbOps.initialize()
     .then(envOps.initialize)
     .then(sentryOps.initialize)
     .then(probotOps.initialize)
@@ -38,6 +38,10 @@ dbOps.initialize()
 
         startPruner();
 
+        return server;
+
     }).catch(error => {
         logger.error(error);
     });
+
+export default async () => await serverHandle;