+++ libs/api/src/v0/clusters.ts
@@ -1,0 +1,121 @@
+import { z } from 'zod';
+import { createTRPCRouter, publicProcedure } from '../trpc';
+
+export const activityRouter = createTRPCRouter({
+    getAllocationByDeploymentId: publicProcedure
+        .input(z.object({
+            deploymentId: z.string().uuid()
+        }))
+        .query(async ({ ctx: { prisma }, input: { deploymentId } }) => {
+
+            if (!deploymentId)
+                return [];
+
+            const deployment = await prisma.deployment.findUnique({
+                where: {
+                    id: deploymentId
+                },
+                select: {
+                    application: {
+                        select: {
+                            organisationId: true
+                        }
+                    }
+                }
+            });
+
+            if (!deployment)
+                return [];
+
+            return await prisma.clusterAllocation.findMany({
+                where: {
+                    organisation: {
+                        id: deployment.application.organisationId
+                    },
+                    AND: {
+                        OR: [{
+                            read: true
+                        }, {
+                            write: true
+                        }, {
+                            admin: true
+                        }]
+                    }
+                },
+                include: {
+                    cluster: true
+                },
+                orderBy: {
+                    createdAt: 'desc'
+                }
+            });
+        }),
+    getAllocationByOrganisationId: publicProcedure
+        .input(z.object({
+            organisationId: z.string().uuid()
+        }))
+        .query(async ({ ctx: { prisma }, input: { organisationId } }) => {
+
+            if (!organisationId)
+                return [];
+
+            return await prisma.clusterAllocation.findMany({
+                where: {
+                    organisation: {
+                        id: organisationId
+                    },
+                    AND: {
+                        OR: [{
+                            read: true
+                        }, {
+                            write: true
+                        }, {
+                            admin: true
+                        }]
+                    }
+                },
+                include: {
+                    cluster: true
+                },
+                orderBy: {
+                    createdAt: 'desc'
+                }
+            });
+        }),
+    deleteAllocation: publicProcedure
+        .input(z.object({
+            allocationId: z.string().uuid()
+        }))
+        .mutation(async ({ ctx: { prisma }, input: { allocationId } }) => {
+
+            if (!allocationId)
+                return false;
+
+            await prisma.clusterAllocation.delete({
+                where: {
+                    id: allocationId
+                }
+            });
+
+            return true;
+        }),
+    delete: publicProcedure
+        .input(z.object({
+            clusterId: z.string().uuid()
+        }))
+        .mutation(async ({ ctx: { prisma }, input: { clusterId } }) => {
+
+            if (!clusterId)
+                return false;
+
+            await prisma.cluster.delete({
+                where: {
+                    id: clusterId
+                }
+            });
+
+            return true;
+        })
+});
+
+export default activityRouter;