+++ apps/hubber/CHANGELOG.md
@@ -1,4 +1,27 @@
 # Changelog
+## [0.3.3](https://github.com/klave-network/platform/compare/hubber@0.3.2...hubber@0.3.3) (2024-07-25)
+
+### Dependency Updates
+
+* `db` updated to version `0.3.2`
+* `providers` updated to version `0.3.2`
+* `constants` updated to version `0.3.2`
+* `api` updated to version `0.3.2`
+* `pruner` updated to version `0.3.2`
+## [0.3.2](https://github.com/klave-network/platform/compare/hubber@0.3.1...hubber@0.3.2) (2024-07-24)
+
+### Dependency Updates
+
+* `db` updated to version `0.3.1`
+* `providers` updated to version `0.3.1`
+* `constants` updated to version `0.3.1`
+* `api` updated to version `0.3.1`
+* `pruner` updated to version `0.3.1`
+
+### Bug Fixes
+
+* **hubber:** Sentry sourcemap ESBuild configuration ([d0f2b87](https://github.com/klave-network/platform/commit/d0f2b87c9887a26cf58aeab53ee6a80d17aa9f92))
+
 ## [0.3.1](https://github.com/klave-network/platform/compare/hubber@0.3.0...hubber@0.3.1) (2024-07-03)
 
 ### Dependency Updates